Moved MeshComponent to its own file (comp_mesh.cpp) from gamesys.cpp (#6455)
Did commented code cleanup too
